chore: Refactor distribution
