--- built-in.js
@@ -1,5 +1,0 @@
-import EnhancedTextAreaBuiltIn, { defineElement } from './dist/built-in/index.js';
-
-export default EnhancedTextAreaBuiltIn;
-
-export { defineElement };
