--- src/built-in/index.js
@@ -1,14 +1,0 @@
-import EnhancedTextAreaBuiltInFactory from './factory.js';
-
-export const defineElement = (name = 'enhanced-textarea') => {
-  customElements.define(
-    name,
-    EnhancedTextAreaBuiltIn,
-    { extends: 'textarea' }
-  );
-};
-
-const EnhancedTextAreaBuiltIn =
-  EnhancedTextAreaBuiltInFactory(HTMLTextAreaElement);
-
-export default EnhancedTextAreaBuiltIn;
