+++ src/with-resize-event/WithResizeEvent.js
