+++ src/with-resize-event/WithResizeEvent.spec.js
