+++ with-autoheight.js
@@ -1,0 +1,3 @@
+import WithAutoheight from './dist/with-autoheight/WithAutoheight.js';
+
+export default WithAutoheight;
