+++ with-resize-event.js
@@ -1,0 +1,3 @@
+import WithResizeEvent from './dist/with-resize-event/WithResizeEvent.js';
+
+export default WithResizeEvent;
